ugc_service/requirements.txt
aioredis==2.0.1
anyio==3.7.0
async-timeout==4.0.2
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
exceptiongroup==1.1.2
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
orjson==3.8.10
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.7.1
uvicorn==0.21.1
